fix: resolve SonarCloud reliability bugs in local dataset lookup
- Replace volatile field with AtomicReference in CsvNameIndexManifestLoader
- Cast char to long before subtraction in CsvPostingDecoder
- Replace unconditional return in for-loop with iterator in LocalDatasetBucketScanner

+++ src/main/java/ch/uzh/ifi/hase/soprafs26/service/localdatasetlookup/CsvPostingDecoder.java
@@ -35,7 +35,7 @@ public class CsvPostingDecoder {
       boolean sawDigit = false;
       while (index < length && Character.isDigit(encodedProductIndices.charAt(index))) {
         sawDigit = true;
-        delta = Math.addExact(Math.multiplyExact(delta, 10L), encodedProductIndices.charAt(index) - '0');
+        delta = Math.addExact(Math.multiplyExact(delta, 10L), (long) encodedProductIndices.charAt(index) - '0');
         index += 1;
       }
